feat(payment system): enhancement
